Moved Tika parser to external #114
